Email switching

+++ database/reset_database.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "`hostname`" == 'lsf.berea.edu' ]; then
+	echo "DO NOT RUN THIS SCRIPT ON PRODUCTION UNLESS YOU REALLY REALLY KNOW WHAT YOU ARE DOING"
+	exit 1
+fi
+
 echo "Dropping databases"
 mysql -u root -proot --execute="DROP DATABASE \`lsf\`; DROP USER 'lsf_user';"
 mysql -u root -proot --execute="DROP DATABASE \`UTE\`; DROP USER 'tracy_user';"
@@ -20,6 +25,8 @@ rm -rf lsf_migrations
 rm -rf tracy_migrations
 rm -rf migrations.json
 
-# TODO only if we want to add fake data
-echo "Adding demo data"
-python demo_data.py
+# Adding data we need in all environments
+python3 base_data.py
+
+# Adding fake data (we don't want this in a production environment)
+python3 demo_data.py
